requirements.txt
Flask~=2.2.0
Flask-Cors==3.0.8
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.3
alembic==1.11.1
onetimepass==1.0.1
openpyxl==3.0.7
py_pdf_parser~=0.13